Add costing of global writes
Previously we were not tracking or charging for writing to or expanding global memory. This diff now adds in costing for this operation. The cost is calculated as:
1. Charge the cost of the operation as though it were a local ref.
2. If the ref is global, charge `GLOBAL_MEMORY_PER_BYTE_COST` per byte over the value that was previously there (i.e., charge by the amount that memory is expanded).
3. Charge `GLOBAL_MEMORY_PER_BYTE_COST` per byte written to storage. This is meant to cover the cost of the IO operations themselves (and number of iops allowed on the SSD before failure).

We don't special case `ReadRef`s based upon whether or not the reference is global since we charge up front for all data that is pulled into the local cache. I explain this a bit more in the inline comment for the `ReadRef` rule.