Enforce max vec length by bounding binary size
The static analysis tool MIRAI (https://github.com/facebookexperimental/MIRAI) was run over the VM source code. One of MIRAI's functions is to check for potential violations of preconditions on standard library functions.

In this case, the `binary` vector size was found to not bounded so a `push` could fail with a panic if there were too many elements in the vector. We add checks at all locations that push to the binary vector to check for overflow and to bail if an overflow will happen.