[types] fix serialization of Transaction
Fix the serialization of the Transaction type so that it adheres
properly to the LCS spec.

Currently a Transaction is being serialized using a u8 for the enum
variant index whereas the LCS spec dictates that a u32 should be used.
this patch is a breaking change and changes Transaction to use a u32 to
specify its enum variant in its serialized form.

Ref: #1307

Closes: #1372
Approved by: msmouse