[enhancement] Follow-up changes in libra_channel
Summary
Implement Drop trait for Sender and Receiver
Keep track of when the receiver is dropped. The Sender will log crit! whenever it tries to send to a Receiver which has been dropped.
When a Sender gets dropped, we will log a crit! message as well
Related to #1483

Closes: #1490
Approved by: ankushagarwal