[language][move] Tweak message for invalid tokens and add a test
The compiler term "token" may not be understood by end users, and
in this case the error only occurs when an individual character
is not recognized. I.E., it is a character that is never used
anywhere in Move code. Describe this error as "invalid character"
instead of "invalid token".

Closes: #2454
Approved by: tzakian