[move-prover] Prelude templatemization and benchmark support.
- Using now handlebars Rust templates for the prelude. This allows to generate a prelude from the same source via template expansion. The template parameters supported are:
  - `--use-array-theory=true|false`: if set, native equality will be generated from the prelude template. Default: false.
  - `--use-native-equality=true|false`: if set, native equality will be generated. This is automatically set to true on `--use-array-theory=true`.
  - `--stratification-depth=N`: determines the depth of stratification. A custom template helper is used in the prelude to generate stratified functions based on this option. Defaults to 4.
  - `--aggressive_func_inline=true|false`: if set some functions are inlined even if they are large. Defaults to false.
  - `--type-requires=true|false`: if true, prelude procedures require type correctness of their parameters, otherwise not (the generated Boogie uses `free requires`).
- A new option `--bench-repeat=N` has been introduced. This executes Boogie N times on the verification problem, for benchmarking. Defaults to 1.
- The driver prints out timing information about both translation and solving. This is only seen if `--verbose=info` is set. If `--bench-repeat` is set, the average of all runs will be printed.

Currently, array theory does not terminate on libra_account.move, and seems generally twice as slow. Therefore its turned off by default. This PR should make it simpler to play with it and try to find the culprit.

This PR also fixes something that I thought was already fixed before: that the tests in sources/{functional,stdlib} are run using the default flags, and cargo test thus behaves the same as cargo run for the test sources. The default for verification is to only verify public functions. To behave tests as expected under this default, either functions had to be made public, or the pragma `verify = true` had to be added to the source.

Closes: #3818
Approved by: DavidLDill