[network][connmgr] reset both address and backoff on update
When receiving an updated set of address for a peer, we reset their dial
state (if we're already trying to dial them). As a result, we will begin
our next dial attempt from the first address (which might have changed)
and from a fresh backoff (since the current backoff delay might be maxed
out if we can't reach any of their previous addresses).

This change should help to mitigate some of the issues we've been having
with paritions healing very slowly (2-5 min!) when a peer has some
invalid or unreachable addresses in on-chain discovery.

Closes: #4450
Approved by: davidiw