[network][noise] returning EOF when receiving a frame of length 0u16
In our integration of noise, all noise messages are prefixed with a 16-bit length.
Previously, receiving a prefix of 0u16 would make us reset our read_state to ReadState::Init,
which would in turn prompt us to try to read another frame length.

This might seem fine, but it is a weird behavior if you're trying to read a long stream of 0s,
as this read will never end.
I believe this is the only edge-case where a read can never end
(as noise messages are limited in size due to the 16-bit length constraint),
and so I propose that we EOF if we ever receive a frame of length 0.
I don't see any reason why a peer would ever do that to us, why would they ;_;

Closes: #5655