Remove duplicate code for aggregating vote power
The aggregation of total vote power was implemented slightly differently
in two areas of the `validator_verifier.rs` file.

This commit refactors those implementations into a single function and
adds tests for the case in which a validator increases their voting
power to a value large enough to cause an integer overflow.

Closes: #5693