logger: introduce new log macros
Introduce a new set of log macros (error, warn, info, debug, trace) and
replace the existing ones. The intention is for this new set of macros
to eventually replace the `sl_*` variants and unify our two logging
systems.

Since this replaces the older log macros, this patch also removed the
auto structuring of textual log messages. It was determined that this
auto structuring ended up causing trouble due to the ability to
unintentionally create excessively large log lines.

The new macros added in this patch have the following API:

    ```
    info!(
        key = value, // Where value implements Serialize
        schema, // Where schema implements Schema
        "formatted {:?} message", var // Formatted log message
    );
    ```