[move-prover] Copy Propagation and Transformation Pipeline.
This PR refactors the copy propagation algorithm in `reaching_defs_analysis.rs` and applies it to all bytecode. Previously references where skipped for copy propagation because of some issues which are now fixed.

The PR also removes the use of the `$tmp` variable in the generated bytecode, which is obsolete since the new memory model.

Overall goal was to make code more readable in preparation for understanding and changing the pack/unpack semantics. As a side effect, code may have become some 20% smaller, as well speed seem to have increased by 10-15%.Unfortunately, the butterfly effect required to turn verification off for `staple_lbr`.

*Changes in the algorithm*: In the new version, elimination of dead assignments is left to the `livevar_analysis.rs`, which did already some of this work. The new version also uses the reaching definitions at each program point instead of the exit points; the later would be only correct if we had single-assignment semantics.

*Changes in pipepline tools*: For debugging, a few changes where made: (a) there is now a flag `--dump-bytecode` which causes the prover to dump the intermediate bytecode for each step in the pipeline (b) the format of the bytecode output has been (attempted) to improve (c) some assertions have been added to borrow analysis for debugging. Moreover, the `pipeline.run()` algorithm has been refined that it actually again works breadth first, so processors of functions can access the result of the most recent step in the pipeline for other functions (not used because we do not yet do intra-function analysis and transformation, but might be handy later).

Closes: #5927