[language] follow up to latest prologue/epilogue changes
PR #6727 added 2 checks to the prologue and 1 check to the epilogue.
This PR updates the Move Adapter specification document to match
that change and also fixes a number of minor problems and
inconsistencies:

- Add a new BAD_TRANSACTION_FEE_CURRENCY error and status code
for the case where the transaction fee currency has not been
registered in the TransactionFee module. Otherwise, the status
is reported as INSUFFICIENT_BALANCE_FOR_TRANSACTION_FEE, which
does not make sense in this case that has nothing to do with
any balance.

- Reorder the check in the prologue for BAD_TRANSACTION_FEE_CURRENCY
so that it comes before checking the sender's balance. If both
checks would fail, we should report this one instead of giving
the misleading impression that the sender could fix the problem
by increasing their balance.

- Reorder the check for PROLOGUE_ESEQUENCE_NUMBER_TOO_BIG so that
it does not come after the "too new" sequence number check. The
latter must be the last check in the prologue, since it is
non-fatal during the initial validation (but not during execution).

- Reorder the prologue specification to match the order in the
code.

- Remove the ESEQUENCE_NUMBER error code. It describes the same
situation as the new PROLOGUE_ESEQUENCE_NUMBER_TOO_BIG that was
added in #6727 and we do not need both. (It was also being used
in a redundant check in the epilogue that the txn sequence number
matches the sending account, but the prologue's version of that
check returns PROLOGUE_ESEQUENCE_NUMBER_TOO_NEW and the epilogue
should be consistent with that. Due to the prologue check, I do
not believe this is a breaking change, because it is impossible
to hit this error in the epilogue.)

- Rename the PROLOGUE_INVALID_WRITESET_SENDER constant to
"PROLOGUE_EINVALID_WRITESET_SENDER", matching all the other
prologue error codes that have a "PROLOGUE_E" prefix.

- Remove the EPILOGUE_INVALID_WRITESET_SENDER constant that was
added in #6727. We do not need separate error codes for the
epilogue and should use the existing prologue error code for this.

- Fix some inaccurate comments.

Closes: #7063